+++ src/include/tlct/config/arrange/offset.hpp
@@ -25,7 +25,7 @@ public:
     TLCT_API inline OffsetArrange& operator=(const OffsetArrange& rhs) noexcept = default;
     TLCT_API inline OffsetArrange(OffsetArrange&& rhs) noexcept = default;
     TLCT_API inline OffsetArrange& operator=(OffsetArrange&& rhs) noexcept = default;
-    TLCT_API inline OffsetArrange(cv::Size imgsize, double diameter, bool direction, cv::Point2d offset) noexcept;
+    TLCT_API inline OffsetArrange(cv::Size imgsize, float diameter, bool direction, cv::Point2f offset) noexcept;
 
     // Initialize from
     [[nodiscard]] TLCT_API static inline OffsetArrange fromCfgMap(const ConfigMap& map);
@@ -37,26 +37,26 @@ public:
     [[nodiscard]] TLCT_API inline int getImgWidth() const noexcept { return imgsize_.width; };
     [[nodiscard]] TLCT_API inline int getImgHeight() const noexcept { return imgsize_.height; };
     [[nodiscard]] TLCT_API inline cv::Size getImgSize() const noexcept { return imgsize_; };
-    [[nodiscard]] TLCT_API inline double getDiameter() const noexcept { return diameter_; };
-    [[nodiscard]] TLCT_API inline double getRadius() const noexcept { return radius_; };
+    [[nodiscard]] TLCT_API inline float getDiameter() const noexcept { return diameter_; };
+    [[nodiscard]] TLCT_API inline float getRadius() const noexcept { return radius_; };
     [[nodiscard]] TLCT_API inline bool getDirection() const noexcept { return direction_; };
     [[nodiscard]] TLCT_API inline int getUpsample() const noexcept { return upsample_; };
     [[nodiscard]] TLCT_API inline int getMIRows() const noexcept { return mirows_; };
     [[nodiscard]] TLCT_API inline int getMICols(const int row) const noexcept { return micols_[row % micols_.size()]; };
     [[nodiscard]] TLCT_API inline int getMIMaxCols() const noexcept { return std::max(micols_[0], micols_[1]); };
     [[nodiscard]] TLCT_API inline int getMIMinCols() const noexcept { return std::min(micols_[0], micols_[1]); };
-    [[nodiscard]] TLCT_API inline cv::Point2d getMICenter(int row, int col) const noexcept;
-    [[nodiscard]] TLCT_API inline cv::Point2d getMICenter(cv::Point index) const noexcept;
+    [[nodiscard]] TLCT_API inline cv::Point2f getMICenter(int row, int col) const noexcept;
+    [[nodiscard]] TLCT_API inline cv::Point2f getMICenter(cv::Point index) const noexcept;
     [[nodiscard]] TLCT_API inline bool isOutShift() const noexcept { return is_out_shift_; };
 
 private:
     cv::Size imgsize_;
-    double diameter_;
-    double radius_;
+    float diameter_;
+    float radius_;
     bool direction_;
-    cv::Point2d left_top_;
-    double x_unit_shift_;
-    double y_unit_shift_;
+    cv::Point2f left_top_;
+    float x_unit_shift_;
+    float y_unit_shift_;
     int mirows_;
     TMiCols micols_;
     int upsample_;
@@ -66,9 +66,9 @@ private:
 OffsetArrange OffsetArrange::fromCfgMap(const ConfigMap& map)
 {
     cv::Size imgsize{map.get<"LensletWidth", int>(), map.get<"LensletHeight", int>()};
-    const double diameter = map.get<"MIDiameter", double>();
+    const float diameter = map.get<"MIDiameter", float>();
     const bool direction = map.get_or<"MLADirection">(false);
-    const cv::Point2d offset = {map.get<"CentralMIOffsetX", double>(), map.get<"CentralMIOffsetY", double>()};
+    const cv::Point2f offset = {map.get<"CentralMIOffsetX", float>(), map.get<"CentralMIOffsetY", float>()};
 
     return {imgsize, diameter, direction, offset};
 }
@@ -86,21 +86,21 @@ OffsetArrange& OffsetArrange::upsample(int factor) noexcept
     return *this;
 }
 
-cv::Point2d OffsetArrange::getMICenter(int row, int col) const noexcept
+cv::Point2f OffsetArrange::getMICenter(int row, int col) const noexcept
 {
-    cv::Point2d center{left_top_.x + x_unit_shift_ * col, left_top_.y + y_unit_shift_ * row};
+    cv::Point2f center{left_top_.x + x_unit_shift_ * col, left_top_.y + y_unit_shift_ * row};
     if (row % 2 == 1) {
         center.x -= x_unit_shift_ / 2.0 * _hp::sgn(isOutShift());
     }
     return center;
 }
 
-cv::Point2d OffsetArrange::getMICenter(cv::Point index) const noexcept { return getMICenter(index.y, index.x); }
+cv::Point2f OffsetArrange::getMICenter(cv::Point index) const noexcept { return getMICenter(index.y, index.x); }
 
-OffsetArrange::OffsetArrange(cv::Size imgsize, double diameter, bool direction, cv::Point2d offset) noexcept
-    : diameter_(diameter), radius_(diameter / 2.0), direction_(direction), upsample_(1)
+OffsetArrange::OffsetArrange(cv::Size imgsize, float diameter, bool direction, cv::Point2f offset) noexcept
+    : diameter_(diameter), radius_(diameter / 2.0f), direction_(direction), upsample_(1)
 {
-    cv::Point2d center_mi{imgsize.width / 2.0 + offset.x, imgsize.height / 2.0 - offset.y};
+    cv::Point2f center_mi{imgsize.width / 2.0f + offset.x, imgsize.height / 2.0f - offset.y};
 
     if (getDirection()) {
         std::swap(imgsize.height, imgsize.width);
@@ -113,7 +113,7 @@ OffsetArrange::OffsetArrange(cv::Size imgsize, double diameter, bool direction, 
     const int center_mi_xidx = (int)((center_mi.x - radius_) / x_unit_shift_);
     const int center_mi_yidx = (int)((center_mi.y - radius_) / y_unit_shift_);
 
-    const double left_x = center_mi.x - x_unit_shift_ * center_mi_xidx;
+    const float left_x = center_mi.x - x_unit_shift_ * center_mi_xidx;
     if (center_mi_yidx % 2 == 0) {
         left_top_.x = left_x;
         if (left_top_.x > diameter) {
@@ -132,10 +132,10 @@ OffsetArrange::OffsetArrange(cv::Size imgsize, double diameter, bool direction, 
     }
     left_top_.y = center_mi.y - std::floor((center_mi.y - y_unit_shift_ / 2.0) / y_unit_shift_) * y_unit_shift_;
 
-    const double mi_1_0_x = left_top_.x - x_unit_shift_ / 2.0 * _hp::sgn(is_out_shift_);
-    micols_[0] = (int)(((double)imgsize.width - left_top_.x - x_unit_shift_ / 2.0) / x_unit_shift_) + 1;
-    micols_[1] = (int)(((double)imgsize.width - mi_1_0_x - x_unit_shift_ / 2.0) / x_unit_shift_) + 1;
-    mirows_ = (int)(((double)imgsize.height - left_top_.y - y_unit_shift_ / 2.0) / y_unit_shift_) + 1;
+    const float mi_1_0_x = left_top_.x - x_unit_shift_ / 2.0 * _hp::sgn(is_out_shift_);
+    micols_[0] = (int)(((float)imgsize.width - left_top_.x - x_unit_shift_ / 2.0) / x_unit_shift_) + 1;
+    micols_[1] = (int)(((float)imgsize.width - mi_1_0_x - x_unit_shift_ / 2.0) / x_unit_shift_) + 1;
+    mirows_ = (int)(((float)imgsize.height - left_top_.y - y_unit_shift_ / 2.0) / y_unit_shift_) + 1;
 }
 
 } // namespace tlct::_cfg
