+++ src/include/tlct/config/common/cli.hpp
@@ -38,15 +38,15 @@ namespace fs = std::filesystem;
         .default_value(1);
     parser->add_argument("--psizeInflate")
         .help("the extracted patch will be inflated by this scale")
-        .scan<'g', double>()
+        .scan<'g', float>()
         .default_value(2.15);
     parser->add_argument("--viewShiftRange")
         .help("reserve `viewShiftRange*diameter` for view shifting")
-        .scan<'g', double>()
+        .scan<'g', float>()
         .default_value(0.1);
     parser->add_argument("--patternSize")
         .help("the size of matching pattern will be `patternSize*diameter`")
-        .scan<'g', double>()
+        .scan<'g', float>()
         .default_value(0.3);
     parser->add_argument("--psizeShortcutThre")
         .help("if the difference bit of dhash of MI is smaller than this value, then use the prev. patch size")
@@ -70,7 +70,7 @@ struct CliConfig {
     };
 
     struct Convert {
-        inline Convert(int views, int upsample, double psize_inflate, double view_shift_range, double pattern_size,
+        inline Convert(int views, int upsample, float psize_inflate, float view_shift_range, float pattern_size,
                        int psize_shortcut_threshold) noexcept
             : views(views), upsample(upsample), psize_inflate(psize_inflate), view_shift_range(view_shift_range),
               max_psize((1.0 - view_shift_range) / psize_inflate), pattern_size(pattern_size),
@@ -78,10 +78,10 @@ struct CliConfig {
 
         int views;
         int upsample;
-        double psize_inflate;
-        double view_shift_range;
-        double max_psize;
-        double pattern_size;
+        float psize_inflate;
+        float view_shift_range;
+        float max_psize;
+        float pattern_size;
         int psize_shortcut_threshold;
     };
 
@@ -97,9 +97,9 @@ CliConfig CliConfig::fromParser(const argparse::ArgumentParser& parser)
 {
     auto path = CliConfig::Path{parser.get<std::string>("--src"), parser.get<std::string>("--dst")};
     auto range = CliConfig::Range{parser.get<int>("--begin"), parser.get<int>("--end")};
-    auto convert = CliConfig::Convert{parser.get<int>("--views"),           parser.get<int>("--upsample"),
-                                      parser.get<double>("--psizeInflate"), parser.get<double>("--viewShiftRange"),
-                                      parser.get<double>("--patternSize"),  parser.get<int>("--psizeShortcutThre")};
+    auto convert = CliConfig::Convert{parser.get<int>("--views"),          parser.get<int>("--upsample"),
+                                      parser.get<float>("--psizeInflate"), parser.get<float>("--viewShiftRange"),
+                                      parser.get<float>("--patternSize"),  parser.get<int>("--psizeShortcutThre")};
 
     return {std::move(path), std::move(range), std::move(convert)};
 }
